Made shit work under VS17

+++ game/game/textbox.cpp
@@ -7,7 +7,7 @@ textbox::textbox(std::string string, std::string filename, sf::RenderWindow & wi
 {
     texture.loadFromFile("textbox.png");
     texture2.loadFromFile(filename);
-    position = sf::Vector2f(0, (window.getSize().y*0.75));
+    position = sf::Vector2f(0.0f, (static_cast<float>(window.getSize().y*0.75)));
     box_background.setTexture(&texture);
     head_box.setTexture(&texture2);
     build_box();
@@ -17,10 +17,10 @@ void textbox::build_box() {
     box_background.setPosition(position);
     box_background.setSize(sf::Vector2f(float(window.getSize().x), float(window.getSize().y*0.2) ));
 
-    float posx = position.x *1.05;
-    float posy = position.y *1.05;
-    head_box.setPosition(sf::Vector2f(posx+15,posy*0.98));
-    head_box.setSize(sf::Vector2f(box_background.getSize().x*0.1, box_background.getSize().y*0.80));
+    float posx = static_cast<float>(position.x *1.05);
+    float posy = static_cast<float>(position.y *1.05);
+    head_box.setPosition(sf::Vector2f(posx+15.0f,posy*0.98f));
+    head_box.setSize(sf::Vector2f(box_background.getSize().x*0.1f, box_background.getSize().y*0.80f));
 
     font.loadFromFile("ArchitectsDaughter.ttf");
     text.setFont(font);
@@ -32,7 +32,7 @@ void textbox::build_box() {
 
     sf::Vector2f positionbox = sf::Vector2f(box_background.getPosition());
     float textoffset = ((positionbox.x)+300);
-    text.setPosition((textoffset),(positionbox.y*1.1));
+    text.setPosition((textoffset),(positionbox.y*1.1f));
 }
